refactor support classes
Extract base class 'TestDir' from 'TestRepo' and 'TestSourceDir' classes.
Update 'TestRepo' and 'TestSourceDir' to inherit from 'TestDir'.
