+++ features/support/test_source_dir.rb
@@ -1,20 +1,5 @@
-class TestSourceDir
-  attr_reader :path
-
-  def initialize(path = 'features/fixtures/source')
-    @path = path
-  end
-
-  def nuke
-    FileUtils.rm_rf(@path) if File.directory?(@path)
-  end
+class TestSourceDir < TestDir
   def init
     FileUtils.mkdir(@path)
-  end
-  def create_file(fname, msg)
-    File.open(File.join(@path, fname), 'w') {|f| f.puts(msg)}
-  end
-  def create_sub_directory(rel_path)
-    FileUtils.mkdir_p(File.join(@path, rel_path))
   end
 end
