+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require_relative '../lib/git_transactor'
+require_relative '../features/support/test_dir'
 require_relative '../features/support/test_repo'
 require_relative '../features/support/test_source_dir'
 require_relative '../features/support/test_queue'
